integrated device-independent protocol version of splicer.c

+++ trunk/remotecrypto/splicer.c
@@ -23,11 +23,13 @@
 
    program to process the saved results in a type-3 file on side A together
    with a type-4 file from the coincidence/sifting unit on side B. Result is a
-   reduced type-3 stream which should contain the raw key information.
+   reduced type-3 stream which should contain the raw key information. This
+   code integrates BB84 and Ekert protocols. 
 
    usage:
       splicer [-i type-3 streamfile] | [-d big type-3 directory]
               [-I type-4 streamfile] | [-D type-4 directory] | [-S socket4]
+	      [-b type-5 bellfile]   | [-B type-5 dell directory]
 	      [-o out type-3 streamfile] | [-f out type-3 directory] | 
 	         [-z socket3]
               [-e start epoch] [-q epochnumber] | [-E cmdpipeline]
@@ -52,6 +54,10 @@
    -o outfile3:     Outfile name for type-3 sifted key packets. This option
                     saves all type-3 packets into a the file/FIFO named
 		    outfile3.
+   -b outfile5:     target file name for a test file for carrying out a
+                    Ekert-type protocol.
+   -B outdir5:      All type-5 packets are saved in this directory, file
+                    name is the epoch in hex.
    -f outdir3:      All type-3 sifted key packets are saved into the directory
                     outdir4, with the file name being the epoch (filling zero
 		    expanded) in hex. Filename is not padded at end.
@@ -72,6 +78,9 @@
 		       for those entries marked in stream 4
 		    1: selects basebits from stream 3in which are marked
 		       in stream4
+		    2: same as mode 0
+		    3: device-independent protocol, this side has 6 detectors
+		    4: device-independent proto, this side has 4 detectors
    -q epochunum:    number of epocs to be read. When set to 0, it loops
                     forever; this is the default.
 
@@ -93,6 +102,7 @@ History:
    written first specs 21.8.05 chk
    compiles; 29.8.05chk
    completed -l, -L, -m options 26.11.05chk
+   integrated with E91 protocol branch with -b and -B options 29.7.11chk
 
 ToDo:
    checking -in progress, 
@@ -182,16 +192,17 @@ typedef struct header_4 {/* header for type-4 stream packet */
 typedef struct protocol_details_C { /* used in splicer program */
     int expected3ibits; /* bits expected from local bit source */
     int expected4ibits; /* bits expected from remote bit source */
-    int transmittedbits; /* bits which should end up in target file */
+    int transmittedbits; /* bits which should end up in target key file */
+    int testfilebits;    /* number of bits in test file */
     int decsize; /* size of an array containing the outpattern3, where
 		    the parameter is the ORed version of the local bits
 		    from stream 3in, and the bits received from stream4in,
 		    shifted by exected3ibits to the left. This size should be
 		    1<<(expected3bits+expected4bits). */
-    void (*filltable)(unsigned int*); /* helper function to fill this array. Has to be
-			        called in the beginning */
+    void (*filltable)(unsigned int*); /* helper function to fill this array.
+					 Has to be called in the beginning */
 } pd_C;
-#define PROTOCOL_MAXINDEX 1
+#define PROTOCOL_MAXINDEX 4
 void FILL_TABLE_PROTO0(unsigned int *t) {
     int i;
     for (i=0;i<256;i++) 
@@ -202,17 +213,60 @@ void FILL_TABLE_PROTO1(unsigned int *t) {
     t[0]=0;t[1]=1;
     return;
 }
+void FILL_TAB_3_4_HELP(unsigned int *t) { /* devindependent common part for 3/4.
+					    argument
+				   is remote response in bits 5-7, and local
+				   detector pattern in bits 0-3. The resulting
+				   word has the stream3/3 data in bits 0-3, for
+				   bell data in the form llrr to make the bell
+				   files on both sides identical. bit 4 is the
+				   ignore indicator, and bit 5 the test pattern
+				   indicator.
+				 */
+    int i;
+    for (i=0;i<128;i++) t[i]=0x10; /* default pattern is ignore */
+    /* all 16 bell test patterns */
+    for (i=0;i<16;i++) t[((i&3)<<4)+(1<<((i>>2)&3))]=0x20+i;
+    return;
+}
+void FILL_TABLE_PROTO3(unsigned int *t) {/* version for local w 6 detectors */
+    FILL_TAB_3_4_HELP(t);   /* fill ignores and test patterns */
+    /* data patterns */
+    t[0x46]=0; t[0x43]=1;     
+    return;
+}
+void FILL_TABLE_PROTO4(unsigned int *t) {/* version for local w 4 detectors */
+    /* data patterns; this is the only difference compared to protocol 3 */
+    FILL_TAB_3_4_HELP(t);   /* fill ignores and test patterns */
+    t[0x42]=0; t[0x41]=1;
+    return;
+}
+
+
 struct protocol_details_C proto_table[] = {
     {/* service protocol. emits all bits in stream 3i and 4i into the outword.
 	the 3i bits end up in bits 4..7, the 4i inbits in bits0..3. This gives
 	the same service files on both sides. It expects 4 bits from stream 3i
 	and 4i. */
-	4,4,8,256,
+	4,4,8,0,256,
 	&FILL_TABLE_PROTO0},
     {/* standard BB84 protocol. expects 1 bit from stream 3i, and nothing from
       stream 4i. reflects the inbit to the outbit. */
-	1,0,1,2,
+	1,0,1,0,2,
 	&FILL_TABLE_PROTO1},
+    {/* duplicated service protocol. same as proto 0. */
+	4,4,8,0,256,
+	&FILL_TABLE_PROTO0},
+    {/* device-independent protocol with 6 detectors on local side. Expects
+	4 full data bits in local file, and 3 data bits in remote data item.
+     */
+	4,3,1,4,128,
+	&FILL_TABLE_PROTO3},
+    {/* device-independent protocol with 4 detectors on local side. Expects
+	4 full data bits in local file, and 3 data bits in remote data item.
+     */
+	4,3,1,4,128,
+	&FILL_TABLE_PROTO4},
 };
 
 
@@ -263,15 +317,22 @@ char *errormessage[] = {
   "cannot stat command pipeline", /* 40 */
   "cmdpipe is not a FIFO",
   "cannot read current epoch from pipe",
-
+  "Error reading file/directory name for type-5 output packets.",
+  "duplicate definition of type-5 outfile.",
+  "Cannot malloc stream-5o buffer.", /* 45 */
+  "error opening target stream 5", 
+  "Cannot write header of stream-5",
+  "Error writing data to stream-3",
 };
 int emsg(int code) {
   fprintf(stderr,"%s\n",errormessage[code]);
   return code;
 };
 
-int openmode[3] = {O_RDONLY,O_RDONLY, /* modes for input streams */
-		   O_WRONLY | O_TRUNC | O_CREAT /* outstream 3 */
+int openmode[5] = {O_RDONLY,O_RDONLY, /* modes for input streams */
+		   O_WRONLY | O_TRUNC | O_CREAT, /* outstream 3 */
+		   0, /* don't care */
+		   O_WRONLY | O_TRUNC | O_CREAT /* outstream 5 */
 };
 
 #define FILE_PERMISSIONS 0644  /* for all output files */
@@ -279,20 +340,22 @@ int openmode[3] = {O_RDONLY,O_RDONLY, /* modes for input streams */
 /* global variables for IO handling */
 int verbosity_level = DEFAULT_VERBOSITY;
 /* handle index: 0: stream3 input, 1: stream4 input, 2: stream3output,
-   4: cmdpipeline input */
-char fname[4][FNAMELENGTH]={"","","",""}; /* stream files / cmd pipeline */
-char ffnam[3][FNAMELENGTH+10]; /* combined dir/filenames */
-int typemode[3]={0,0,0}; /* 0: no mode defined. other types:
+   4: cmdpipeline input, 5: test file output */
+char fname[5][FNAMELENGTH]={"","","","",""}; /* stream files / cmd pipeline */
+char ffnam[5][FNAMELENGTH+10]; /* combined dir/filenames */
+int typemode[5]={0,0,0,0,0}; /* 0: no mode defined. other types:
 			    1: single file, 2: directory save, ... */
 int killmode[2] = {DEFAULT_KILLMODE3,
 		   DEFAULT_KILLMODE4 }; /* if !=1, delete infile after use */
 FILE* loghandle[3]; /* index 0: cnsmd t3, 1: cnsmd t4, 2: made rawk */
 char logfname[3][FNAMELENGTH]={"","",""};
-int handle[3]; /* global handles for packet streams */
+int handle[5]; /* global handles for packet streams */
 unsigned int current_ep;
 struct header_3 head3i; /* infile header */
 struct header_4 head4i; /* infile header */
 struct header_3 head3o; /* outfile header */
+struct header_3 head5o; /* outfile header for test file */
+
 int expected3bits, expected4bits;
 
 int get_stream_3(void *buffer, int handle, int maxsize,
@@ -376,7 +439,7 @@ int main (int argc, char *argv[]) {
     unsigned int epochnumber = DEFAULT_EPOCHNUMBER; /* # of epochs to read  */
     int proto_index = DEFAULT_PROTOCOL; /* defines which proto is used */
     long long int se_in; /* for entering startepoch both in hex and decimal */
-    char *buffer3i, *buffer4i, *buffer3o; /* stream buffers */
+    char *buffer3i, *buffer4i, *buffer3o, *buffer5o; /* stream buffers */
 
     int realsize4i; /* for processing stream 4 */
     unsigned int *pointer4;
@@ -384,17 +447,23 @@ int main (int argc, char *argv[]) {
     unsigned int readword4,emergency_break,pattern4,patternmask4,diff4_bitmask;
     int resbits4;
     int type4bitwidth,type4datawidth,processed_4events,bitstoread4;
-    
+    int processed_testevents; /* for type 5 events */
+    int processed_keyevents; /* for type 3 events */
+
     int k3i,n3i,type3ibits,u3i;
     unsigned int pattern3i, patternmask3i;
     unsigned int *pointer3i;
     int p3mask1[32],p3sh1[32],p3mask2[32],p3sh2[32],p3sh0[32],p3dec[32];
     int stream3imaxindex; /* for input consistency check */
 
-    unsigned int pattern3o;
-    unsigned int sendword3;
-    int resbits3o,type3odatawidth,index3o;
-    unsigned int *outbuf3o, *o3_lookup;
+    int mostbits;
+    unsigned int decpattern;
+    unsigned int pattern3mask, pattern5mask, ignorepatternmask,testbitmask;
+    unsigned int pattern5o,pattern3o;
+    unsigned int sendword3, sendword5;
+    int resbits3o, resbits5o, type3odatawidth, index3o, index5o;
+    int type5odatawidth;
+    unsigned int *outbuf3o, *lookup_table, *outbuf5o;
     
     int opt,i,j,retval;
     unsigned int ju; /* wherever it is needed */
@@ -404,7 +473,7 @@ int main (int argc, char *argv[]) {
     struct stat cmdstat; /* for probing pipe */
 
     opterr=0; /* be quiet when there are no options */
-    while ((opt=getopt(argc, argv, "V:i:d:I:D:o:f:e:q:p:kKl:E:L:m:")) != EOF) {
+    while ((opt=getopt(argc, argv, "V:i:d:I:D:o:f:e:q:p:kKl:E:L:m:b:B:")) != EOF) {
 	i=0; /* for setinf names/modes commonly */
 	switch (opt) {
 	    case 'V': /* set verbosity level */
@@ -425,6 +494,13 @@ int main (int argc, char *argv[]) {
 		fname[j][FNAMELENGTH-1]=0;   /* security termination */
 		if (typemode[j]) return -emsg(5+j); /* already defined mode */
 		typemode[j]=(i&4?2:1);
+		break;
+	    case 'B': i++; /* stream 5 directory */
+	    case 'b':   /* stream 5 outfile */
+		if (1!=sscanf(optarg,FNAMFORMAT,fname[4])) return -emsg(43);
+		fname[4][FNAMELENGTH-1]=0;   /* security termination */
+		if (typemode[4]) return -emsg(44); /* already defined mode */
+		typemode[4]=(i&1?2:1);
 		break;
 	    case 'e': /* read startepoch */
 		if (1!=sscanf(optarg,"%lli",&se_in)) return -emsg(8);
@@ -466,16 +542,26 @@ int main (int argc, char *argv[]) {
     if (!(buffer3i=(char*)malloc(RAW3i_SIZE))) return -emsg(13);
     if (!(buffer4i=(char*)malloc(RAW4i_SIZE))) return -emsg(14);
     if (!(buffer3o=(char*)malloc(RAW3o_SIZE))) return -emsg(15);
+    if (!(buffer5o=(char*)malloc(RAW3o_SIZE))) return -emsg(45);
 
     /* prepare protocol specific stuff */
+    mostbits = proto_table[proto_index].transmittedbits;
+    if (mostbits<proto_table[proto_index].testfilebits) 
+	mostbits=proto_table[proto_index].testfilebits; /* max length */
     type3odatawidth=proto_table[proto_index].transmittedbits;
+    type5odatawidth=proto_table[proto_index].testfilebits;
+    pattern3mask = (1<<type3odatawidth)-1;
+    pattern5mask = (1<<type5odatawidth)-1;
+    ignorepatternmask = (1<<mostbits); /* for extracting the ignore tag */
+    testbitmask = ignorepatternmask<<1; /* for deciding which file to fill */
+    
     expected3bits=proto_table[proto_index].expected3ibits;
     expected4bits=proto_table[proto_index].expected4ibits;
     /* fill outword table */
-    o3_lookup=(unsigned int *)malloc(sizeof(int)*
+    lookup_table=(unsigned int *)malloc(sizeof(int)*
 			    proto_table[proto_index].decsize);
-    if (!o3_lookup ) return -emsg(36);
-    proto_table[proto_index].filltable(o3_lookup);
+    if (!lookup_table ) return -emsg(36);
+    proto_table[proto_index].filltable(lookup_table);
 
     type3ibits=expected3bits;
     patternmask3i=(1<<type3ibits)-1;
@@ -507,7 +593,10 @@ int main (int argc, char *argv[]) {
 		break;
 	}
     }
-
+    if (typemode[4]==1) { /* open steam-5 output file */
+		handle[4]=open(fname[4],openmode[4],FILE_PERMISSIONS);
+		if(-1==handle[4]) return -emsg(46);
+    }	
 
     if (cmdmode) { /* load initial epoch */
 	if (1!=fscanf(cmdhandle,"%x",&current_ep)) return -emsg(42);
@@ -541,6 +630,13 @@ int main (int argc, char *argv[]) {
 	head3o.length=0;
 	head3o.bitsperentry=type3odatawidth;
 
+	/* prepare outstream-5 */
+	head5o.tag=(head4i.tag==TYPE_4_TAG_U)?
+	    TYPE_3_TAG_U:TYPE_3_TAG; /* same style */
+	head5o.epoc=current_ep;
+	head5o.length=0;
+	head5o.bitsperentry=type5odatawidth;
+
 	/* prepare parsing instream-4 elements */
 	tindex=0; /* initial value */
 	pointer4=(unsigned int *)(buffer4i+sizeof(struct header_4));
@@ -553,10 +649,14 @@ int main (int argc, char *argv[]) {
 	patternmask4=(1<<type4datawidth)-1;
 	diff4_bitmask=(1<<type4bitwidth)-1;
 	processed_4events=0; /* processed events */
+	processed_testevents=0; /* test events */
+	processed_keyevents=0; /* key events */
 
-        sendword3=0;
+        sendword3=0; sendword5=0;
 	outbuf3o=(unsigned int *)buffer3o;
-	resbits3o=32;index3o=0;
+	outbuf5o=(unsigned int *)buffer5o;
+	resbits3o=32;index3o=0;resbits5o=32;index5o=0;
+
 
 	do { /* go through buffer */
 	    if (resbits4>=bitstoread4) {
@@ -602,33 +702,57 @@ int main (int argc, char *argv[]) {
 	    }
 	    /* we now have the pattern in pattern3i */
 
-	    /* type-3 stream filling */
-	    pattern3o=o3_lookup[pattern3i 
+	    /* type-3 and type-5 stream filling */
+	    decpattern=lookup_table[pattern3i 
 				| (pattern4<<type3ibits)]; /* out pattern */
-	    if (resbits3o>=type3odatawidth) {
-		sendword3 |= (pattern3o << (resbits3o-type3odatawidth));
-		resbits3o = resbits3o-type3odatawidth;
-		if (resbits3o==0) { 
-		    outbuf3o[index3o++]=sendword3;
-		    sendword3=0;resbits3o=32;
+	    if (!(decpattern & ignorepatternmask)) { /* don't ignore */
+		if (decpattern & testbitmask) { /* we have a stream-5 event */
+		    pattern5o = decpattern & pattern5mask;
+		    if (resbits5o>=type5odatawidth) {
+			sendword5 |= (pattern5o << (resbits5o-type5odatawidth));
+			resbits5o = resbits5o-type5odatawidth;
+			if (resbits5o==0) { 
+			    outbuf5o[index5o++]=sendword5;
+			    sendword5=0;resbits5o=32;
+			}
+		    } else {
+			resbits5o=type5odatawidth-resbits5o;
+			sendword5 |= (pattern5o >> resbits5o);
+
+			outbuf5o[index5o++]=sendword5;
+			resbits5o=32-resbits5o;
+			sendword5=pattern5o << resbits5o;
+		    }
+		    processed_testevents++;
+		} else { /* we have a stream-3 event */
+		    pattern3o = decpattern & pattern3mask;
+		    if (resbits3o>=type3odatawidth) {
+			sendword3 |= (pattern3o << (resbits3o-type3odatawidth));
+			resbits3o = resbits3o-type3odatawidth;
+			if (resbits3o==0) { 
+			    outbuf3o[index3o++]=sendword3;
+			    sendword3=0;resbits3o=32;
+			}
+		    } else {
+			resbits3o=type3odatawidth-resbits3o;
+			sendword3 |= (pattern3o >> resbits3o);
+			outbuf3o[index3o++]=sendword3;
+			resbits3o=32-resbits3o;
+			sendword3=pattern3o << resbits3o;
+		    }
+		    processed_keyevents++;
 		}
-	    } else {
-		resbits3o=type3odatawidth-resbits3o;
-		sendword3 |= (pattern3o >> resbits3o);
-		outbuf3o[index3o++]=sendword3;
-		resbits3o=32-resbits3o;
-		sendword3=pattern3o << resbits3o;
 	    }
 	    processed_4events++;
 	} while (ju<=emergency_break); /* make sure to eat up rest properly */
 	
-	/* output outstream-3 */
-	head3o.length=processed_4events;
+
+	/* finalize outstream-3 */
+	head3o.length=processed_keyevents;
 	if (resbits3o<32) outbuf3o[index3o++]=sendword3; /* last word */
 	
 	/* eventually open output stream 3 */
 	if (openstream(2,current_ep)) return -emsg(19);
-	
 	/* write header 3 */
 	retval= write(handle[2],&head3o,sizeof(struct header_3));
 	if (retval!=sizeof(struct header_3)) 
@@ -636,16 +760,33 @@ int main (int argc, char *argv[]) {
 	i=index3o*sizeof(unsigned int);
 	retval=write(handle[2],outbuf3o,i);
 	if (retval!=i) return -emsg(33); /* write error buffer */
-	
 	/* eventually close stream 3 */
 	if (typemode[2]==2) close(handle[2]);
 
+	/* finalize outstream-5 */
+	head5o.length=processed_testevents;
+	if (resbits5o<32) outbuf5o[index5o++]=sendword5; /* last word */
+	
+	/* eventually open output stream 5 */
+	if (openstream(4,current_ep)) return -emsg(46);
+	/* eventually write header+data of stream 5 */
+	if (typemode[4]) {
+	    retval= write(handle[4],&head5o,sizeof(struct header_3));
+	    if (retval!=sizeof(struct header_3)) 
+		return -emsg(47); /* write head err */
+	    i=index5o*sizeof(unsigned int);
+	    retval=write(handle[4],outbuf5o,i);
+	    if (retval!=i) return -emsg(48); /* write error buffer */
+	}  
+	/* eventually close stream 3 */
+	if (typemode[4]==2) close(handle[4]);
+	
 	/* eventually remove instreams */
 	for (i=0;i<2;i++) if (killmode[i] && (typemode[i]==2)) {
 	    if (unlink(ffnam[i])) return -emsg(34+i);
 	}
 	
-        /* do logging */
+       /* do logging */
 	for (i=0;i<3;i++) {
 	    if (logfname[i][0]) {
 		switch(verbosity_level) {
@@ -653,19 +794,18 @@ int main (int argc, char *argv[]) {
 			fprintf(loghandle[i],"%08x\n",current_ep);
 			break;
 		    case 1: /* epoch, resulting keybits */
-			fprintf(loghandle[i],"%08x\t%d\n",
-				current_ep,processed_4events);
+			fprintf(loghandle[i],"%08x\t%d\n", current_ep,
+				processed_keyevents);
 			break;
 		    case 2: /* same w text */
 			fprintf(loghandle[i],"epoch: %08x, final events: %d\n",
-				current_ep,processed_4events);
+				current_ep,processed_keyevents);
 			break;
 		}
 		/* fprintf(loghandle[i],"bla\n"); */
 		fflush(loghandle[i]);
 	    }
 	}
-	
 
 	if (cmdmode) { /* wait for next epoch */
 	    if (1!=fscanf(cmdhandle,"%x",&current_ep)) return -emsg(42);
@@ -678,9 +818,10 @@ int main (int argc, char *argv[]) {
     /* evtl. close stream files */
     for (i=1;i<3;i++) 
 	if (1== (typemode[i])) close(handle[i]); /* single file */
+    if (1==typemode[4]) close(handle[4]);
 
     for (i=0;i<3;i++) if (logfname[i][0]) fclose(loghandle[i]);
     /* free buffers */
-    free(buffer3i);free(buffer4i);free(buffer3o);
+    free(buffer3i);free(buffer4i);free(buffer3o);free(buffer5o);
     return 0;
 }
